Add settings, possibility to change server address, refactor UI (#7)
Now there's a settings button, which brings up settings pane.
For now it allows to change SERVER_ADDRES. It was moved to Globals.gd to
allow easy changing it on the fly.

Refactor UI, to emit signal for creating agent. Now when signal is
emitted, AgentCreator in scene.tscn catches it and sends addObject
request.
AgentCreator was moved outside of WorldUpdater, because WorldUpdater
should just recreate the scene and apply deltas. All other logic should
be in separate nodes.