Fix push time race
Sometimes when pushing wood, it would be pushed two times instead of
one. It is due to the fact that pushing is done by applying move action
on the log.
However move action wasn't executed in the same frame as push action.
It was just added to action queue, which created a delay.
That delay was causing a time race.